gh-code-review-agent/pyproject.toml
[project]
name = "gh-code-review-agent"
version = "0.1.0"
description = "An autonomous code review agent"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "agno>=1.8.1",
    "bcrypt>=4.3.0",
    "celery>=5.5.3",
    "fastapi[standard]>=0.116.1",
    "flask>=3.1.2",
    "jinja2>=3.1.6",
    "psycopg2>=2.9.10",
    "pydantic-settings>=2.10.1",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.1",
    "redis>=6.4.0",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.35.0",
]
